Rename mounttable to MountTable to keep file name consistent
